installer & updater

+++ up.sh
@@ -4,9 +4,11 @@ rm -rf /tmp/*-shetup-*
 out=/tmp/shetup.tar.gz
 wget -O $out https://github.com/nu7hatch/shetup/tarball/master
 tar -xzf $out -C /tmp
-cp -rf /tmp/*-shetup-*/src ./setup
+mkdir -p ./setup/
+cp -rf /tmp/*-shetup-*/src/*.bash ./setup/
 
-cat > ./setup.sh <<EOF
+if [ ! -f ./setup.sh ]; then
+    cat > ./setup.sh <<EOF
 #!/bin/bash
 
 . ./setup/core.bash
@@ -15,5 +17,5 @@ check_cmd test
 
 success "Package has been successfully configured. Check README.md for more details!"
 EOF
-
-chmod +x ./setup.sh
+    chmod +x ./setup.sh
+fi
